+++ layouts/search/single.html
@@ -2,7 +2,7 @@
 <!-- Search form -->
 {{ partial "search-form.html" . }}
 
-<ul id="results"></ul>
+<div id="results"></div>
 
 <!-- end -->
 {{ end }}
