+++ package.json
@@ -26,5 +26,8 @@
     "stylelint-config-standard": "^20.0.0",
     "stylelint-config-standard-scss": "^3.0.0",
     "stylelint-prettier": "^2.0.0"
+  },
+  "dependencies": {
+    "preact": "^10.6.4"
   }
 }
